+++ packages/stream_video/lib/src/webrtc/rtc_audio_api/rtc_audio_stub.dart
@@ -3,6 +3,8 @@ import 'package:stream_webrtc_flutter/stream_webrtc_flutter.dart' as rtc;
 /// Signature of the sink the web audio layer reports playback traces to.
 typedef AudioTraceHandler = void Function(String tag, Object? data);
 
+Stream<bool> get audioPlaybackBlockedChanges => throw UnimplementedError();
+
 void setAudioTraceHandler(AudioTraceHandler? handler) {
   throw UnimplementedError();
 }
@@ -22,10 +24,6 @@ Future<void> setSinkId(String id, String deviceId) {
 Future<void> resumeAudioPlayback() {
   throw UnimplementedError();
 }
-
-bool get isAudioPlaybackBlocked => throw UnimplementedError();
-
-bool get isAudioPlaybackStalled => throw UnimplementedError();
 
 bool checkIfAudioOutputChangeSupported() {
   throw UnimplementedError();